# contracts/openapi/backend-v1.yaml
openapi: 3.1.0
info:
  title: TrackMyProps Backend API
  version: 1.5.0
  description: Operational, authenticated identity, owner property, and cash-flow endpoints.
...
paths:
...
  /api/v1/me:
    get:
      operationId: getCurrentUser
      summary: Retrieve the authenticated user
      tags:
        - Identity
      security:
        - supabaseBearer: []
      responses:
        "200":
          description: The identity represented by the verified access token.
          content:
            application/json:
              schema:
                $ref: ../json-schema/identity/current-user.schema.json
              examples:
                owner:
                  value:
                    email: owner@example.com
                    id: e8cf2dbf-463e-485f-880d-cdb828749979
                    name: Owner Example
                    phone: null
        "401":
          description: The access token is missing, invalid, or expired.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "503":
          description: Supabase Auth could not verify the access token.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
  /api/v1/properties:
    get:
      operationId: listProperties
      summary: List one lifecycle group of the authenticated owner's properties
      tags:
        - Properties
      security:
        - supabaseBearer: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
            default: active
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Non-deleted properties in the requested lifecycle group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyList"
        "401":
          description: Authentication is missing or invalid.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "503":
          description: Property persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
    post:
      operationId: createProperty
      summary: Create a currently owned property
      tags:
        - Properties
      security:
        - supabaseBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyCreate"
      responses:
        "201":
          description: The property was created for the authenticated owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "401":
          description: Authentication is missing or invalid.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "422":
          description: Property values failed validation or database constraints.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "503":
          description: Property persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
  /api/v1/properties/{property_id}:
    get:
      operationId: getProperty
      summary: Retrieve one owner-visible property
      tags:
        - Properties
      security:
        - supabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "200":
          description: One non-deleted property visible to the authenticated owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "401":
          description: Authentication is missing or invalid.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "404":
          description: No owner-visible property matched the identifier.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "503":
          description: Property persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
    put:
      operationId: updateProperty
      summary: Replace the editable details of an owner property
      tags:
        - Properties
      security:
        - supabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyCreate"
      responses:
        "200":
          description: The property details were updated without changing lifecycle status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "401":
          description: Authentication is missing or invalid.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "404":
          description: No owner-visible property matched the identifier.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "422":
          description: Property values failed validation or database constraints.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "503":
          description: Property persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
  /api/v1/properties/{property_id}/status:
    patch:
      operationId: updatePropertyStatus
      summary: Archive or restore an owner property
      tags:
        - Properties
      security:
        - supabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyStatusUpdate"
      responses:
        "200":
          description: The property was moved to the requested lifecycle list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "401":
          description: Authentication is missing or invalid.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "404":
          description: No owner-visible property matched the identifier.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "422":
          description: The lifecycle transition was rejected.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "503":
          description: Property persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
...
  /api/v1/income/{item_id}:
    delete:
      operationId: deletePropertyIncome
      summary: Remove an income item from an owner property
      tags:
        - Cash flow
      security:
        - supabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: The income item was removed.
        "401":
          description: Authentication is missing or invalid.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "404":
          description: No owner-visible property or income item matched the identifiers.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "503":
          description: Cash-flow persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
...
  /api/v1/expenses/{item_id}:
    delete:
      operationId: deletePropertyExpense
      summary: Remove an expense item from an owner property
      tags:
        - Cash flow
      security:
        - supabaseBearer: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: The expense item was removed.
        "401":
          description: Authentication is missing or invalid.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "404":
          description: No owner-visible property or expense item matched the identifiers.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
        "503":
          description: Cash-flow persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: ../json-schema/errors/error-response.schema.json
...
components:
  parameters:
    PropertyId:
      name: property_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ItemId:
      name: item_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Property:
      allOf:
        - $ref: "#/components/schemas/PropertyCreate"
        - type: object
          required:
            - property_id
            - owner_user_id
            - status
            - created_at
            - updated_at
          properties:
            property_id:
              type: string
              format: uuid
            owner_user_id:
              type: string
              format: uuid
            status:
              type: string
              enum: [active, sold, archived]
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    PropertyList:
      type: object
      additionalProperties: false
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        pagination:
          $ref: ../json-schema/common/pagination.schema.json
    PropertyStatusUpdate:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          enum: [active, archived]
...
  securitySchemes:
    supabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...